Tweaking physics
AirportSecurity zombie should be the corrent weight, as well as the projectiles. Projectile velocity is still TBD but with the current settings, limbs may or may not come off when shot. Perhaps dismemberment can be handled in code later on, rather than being handled by the physics. That would allow for much more fine tuned behavior.